separate data parsing logic from csv parser
