+++ types/fileDataRecord.ts
@@ -1,0 +1,7 @@
+export interface FileDataRecord {
+  Date: string;
+  From: string;
+  To: string;
+  Narrative: string;
+  Amount: string;
+}
